Revert "xx"
This reverts commit 61b6bf6987a4b80de8e9ed4b9aad246a101cc096.

+++ DwarfEngine/SpriteAnimator.h
@@ -6,30 +6,30 @@
 using namespace std;
 
 struct AnimationItem {
-	int id;
-	int duration;
-	int loop; 
-	int startFrame;
+	Uint16 id;
+	Uint16 duration;
+	Uint16 loop; 
+	Uint16 startFrame;
 	int endFrame;
 	EaseType easing;
 };
 
 struct SpriteSlice {
-	int from;
-	int length;
+	Uint16 from;
+	Uint16 length;
 };
 
 class SpriteAnimator :	public Modifier {
 public:
 	SpriteAnimator(void);
 	~SpriteAnimator(void);
-	void addAnimationItem(int id, int duration, int loop, int startFrame, int endFrame,	EaseType easing);
+	void addAnimationItem(Uint16 id, Uint16 duration, Uint16 loop, Uint16 startFrame, int endFrame,	EaseType easing);
 	virtual void onRender(Uint32 deltaTime);
 	virtual void halt(void);
 	virtual void onStop(void);
 private:
 	AnimationItem* currentAnimationItem;
 	queue<AnimationItem*> animations;
-	int loopCount;
+	Uint16 loopCount;
 };
